#50 refactor: planned-path 参照元を新 _stores/planned-path/ へ切替
- hook (usePlannedPathStore) は新 _stores/planned-path/ (実装/context) を参照
- component/context の DEFAULT_PLANNED_PATH は _stores/planned-path/constants へ直接参照
- intent-store.ts の PlannedPathStore 型は _types 集約 index 経由に統一
  (旧 _stores/planned-path-store.ts・_contexts/planned-path-store-context.tsx 自体はまだ削除しない)
