+++ src/prototypes/time-control/time-control-03/__tests__/intent-store.test.ts
@@ -3,7 +3,7 @@ import { createActorSettingsStore } from '../_stores/actor-settings'
 import { createGameClockStore } from '../_stores/game-clock'
 import { createIntentStore } from '../_stores/intent-store'
 import { createPathStore } from '../_stores/path'
-import { createPlannedPathStore } from '../_stores/planned-path-store'
+import { createPlannedPathStore } from '../_stores/planned-path'
 import { createPositionStore } from '../_stores/position-store'
 
 const setup = () => {
