+++ src/prototypes/time-control/time-control-03/__tests__/planned-path-store.test.ts
@@ -1,4 +1,4 @@
-import { createPlannedPathStore } from '../_stores/planned-path-store'
+import { createPlannedPathStore } from '../_stores/planned-path'
 
 test('未設定 actor の予定経路は空扱い', () => {
   const store = createPlannedPathStore()
