+++ src/prototypes/time-control/time-control-03/_stores/intent-store.ts
@@ -10,8 +10,8 @@ import {
   ActorStore,
   GameClockStore,
   PathStore,
+  PlannedPathStore,
 } from './_types'
-import { PlannedPathStore } from './planned-path-store'
 import { DEFAULT_POSITION, PositionStore } from './position-store'
 
 /** 状態を持たない操作専用 store。移動企図 (経路生成 → path-store への書き込み) のみを行う */
@@ -26,7 +26,7 @@ export type IntentStore = StoreApi<IntentState>
  * @param actorStore 移動速度・tick 発生頻度 (`_stores/actor/store.ts`)
  * @param actorSettingsStore 経路の最低 step 数 (`_stores/actor-settings/store.ts`)
  * @param pathStore 生成した経路の書き込み先 (`_stores/path/store.ts`)
- * @param plannedPathStore 予定位置表示用の経路の書き込み先 (`_stores/planned-path-store.ts`)
+ * @param plannedPathStore 予定位置表示用の経路の書き込み先 (`_stores/planned-path/store.ts`)
  * @param positionStore 企図元の現在位置 (`_stores/position-store.ts`)
  * @param gameClockStore 共通ゲームクロック・履歴ログ (`_stores/game-clock/store.ts`)
  */
